+++ sparsecoding/transforms/__init__.py
@@ -1,7 +1,7 @@
-from .patch import quilt, patchify
+from .patch import quilt, patchify, sample_random_patches
 from .whiten import whiten, compute_whitening_stats
 from .images import whiten_images, compute_image_whitening_stats, WhiteningTransform
 
-__all__ = ['quilt', 'patchify', 'whiten',
+__all__ = ['quilt', 'patchify', 'sample_random_patches', 'whiten',
            'compute_whitening_stats', 'compute_image_whitening_stats',
             'WhiteningTransform', 'whiten_images']
